% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b_estimate_dt}
\alias{b_estimate_dt}
\title{Example filtered bam-readcount allele-frequency table}
\format{
A \code{data.table} with 8,598 rows and 8 columns.
}
\usage{
b_estimate_dt
}
\description{
A filtered allele-frequency table aligned to the genotype dosage table. This
object includes the \code{CHROM_POS_REF} identifier used to match rows across
inputs.
}
\keyword{datasets}
