% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bam_readcount}
\alias{bam_readcount}
\title{Example bam-readcount import}
\format{
A \code{data.table} with 10,000 rows and 10 columns.
}
\usage{
bam_readcount
}
\description{
The first ten columns of a bam-readcount export after import with
\code{data.table::fread()}.
}
\keyword{datasets}
