% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_nonRef_vector_and_genotype_df.R
\name{create_gt_dosages}
\alias{create_gt_dosages}
\title{Convert VCF Genotype Data to Minor Allele Dosages}
\usage{
create_gt_dosages(vcf)
}
\arguments{
\item{vcf}{A VCF object of class \code{vcfR} containing genotype data.}
}
\value{
A data.table with SNP identifiers in column 'rn' and genotype dosages (0, 0.5, or 1) for each donor/sample.
}
\description{
This function processes a VCF (Variant Call Format) object to extract genotype (GT) information and convert it into minor allele dosages. It removes duplicated SNP entries, replaces genotype representations with numeric dosage values, and filters out SNPs with missing data.
}
\details{
\itemize{
\item The function extracts genotype (GT) data from the VCF object.
\item It removes duplicate SNP entries based on chromosome, position, and reference allele.
\item Genotype strings are converted into minor allele dosages:
\itemize{
\item \code{0|0} or \code{0/0} -> \code{0}
\item \code{0|1}, \code{1|0}, \code{0/1}, or \code{1/0} -> \code{0.5}
\item \code{1|1} or \code{1/1} -> \code{1}
}
\item SNPs with missing values are removed.
}
}
\examples{
\dontrun{
library(vcfR)
library(data.table)
library(dplyr)
library(stringr)
vcf = read.vcfR("example.vcf")
dosage_data = create_gt_dosages(vcf)
}

}
