% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_nonRef_vector_and_genotype_df.R
\name{estimate_b_from_bam_readcount}
\alias{estimate_b_from_bam_readcount}
\title{Transform the output of bam-readcount to MAF}
\usage{
estimate_b_from_bam_readcount(bam_readcount, vcf)
}
\arguments{
\item{bam_readcount}{Output of bam-readcount, after reading in first 10 columns, filtering to matching positions
of the VCF (and after passing other user-defined filters such as minimum number of TOTAL_READS, if one so wishes).
Colnames are:
"CHROM": Chromosome.
"POS": Genomic position.
"REF": Reference allele.
"TOTAL_READS": Total number of reads mapping to the variant.
"DEL": Counts matching a deletion ALT allele.
"A":  Counts matching "A" ALT allele.
"C": Counts matching "C" ALT allele.
"G": Counts matching "G" ALT allele.
"T": Counts matching "T" ALT allele.
"N": Counts matching "N" ALT allele (any ambiguous/unknown single variant allele).
"CHROM_POS_REF": union of CHROM, POS, REF columns, separated by "_", for filtering.
"ALT": ALT allele considered for this position.}

\item{vcf}{The VCF genotype file, to get the minor allele and subset to common SNPs. Only takes into account
so far minor alleles which are A,C,T or G (TODO: consider changing this)}
}
\value{
\itemize{
\item A data.table containing per SNP position: the total counts,
counts for the A,C,T and G alleles, the minor allele frequency estimate (b_estimate), the ALT allele,
and the SNP name in chromosome_position_REF_ALT (CHROM_POS_REF) format.
}
}
\description{
Transform the output of bam-readcount to data.table with minor allele frequency (MAF) estimate
Only works with bam-readcount output preprocessed to contain first 10 columns as detailed in vignette
}
