% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_weights.R
\name{is_identical_genotype}
\alias{is_identical_genotype}
\title{Check for Identical Genotypes Across Samples}
\usage{
is_identical_genotype(x, ...)
}
\arguments{
\item{x}{A numeric matrix where rows represent loci and columns represent individuals or samples.}

\item{...}{Additional arguments passed to \code{rowMeans()}.}
}
\value{
A logical vector where each element is \code{TRUE} if all values in the corresponding row are identical, and \code{FALSE} otherwise.
}
\description{
This function determines whether all genotypes in each row of a matrix are identical.
It is a faster alternative to \code{apply(A[,..cols], MARGIN = 1, FUN = function(x) var(x) == 0)},
leveraging row-wise variance calculations.
}
\examples{
mat <- matrix(c(1,1,1, 2,2,2, 1,2,1), nrow = 3, byrow = TRUE)
is_identical_genotype(mat)  # Returns: TRUE, TRUE, FALSE

}
