\name{poodleR-package}
\alias{poodleR}
\alias{poodleR-package}
\title{poodleR: pooled donor deconvolution helpers}
\description{
\code{poodleR} provides helper functions for pooled donor deconvolution from
genotype VCFs and bam-readcount output. The package includes functions to build
minor-allele dosage tables, derive pooled allele-frequency estimates, and solve
the constrained regression used to estimate donor proportions.
}
\seealso{
\code{\link{create_gt_dosages}}, \code{\link{estimate_b_from_bam_readcount}},
\code{\link{calculate_nonRef_vector_and_genotype_df}},
\code{\link{estimate_weights}}
}
\keyword{internal}
