% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gt_dosages.R
\name{create_gt_dosages}
\alias{create_gt_dosages}
\title{Create genotype dosage file (minor allele dosages)}
\usage{
create_gt_dosages(vcf)
}
\arguments{
\item{vcf}{The VCF file (was tested with VCF version 4.2) that has been read in with vcfR.}
}
\value{
A data table with with column numbers equal to number of donors
plus one column with the SNP name in format chromosome_position_refAllele (C_POS_REF)
}
\description{
Transform the VCF with genotypes into a genotype dosage data table.
}
