% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_b_from_bam_readcount.R
\name{estimate_b_from_bam_readcount}
\alias{estimate_b_from_bam_readcount}
\title{Transform the output of bam-readcount to MAF}
\usage{
estimate_b_from_bam_readcount(bam_readcount, vcf)
}
\arguments{
\item{bam_readcount}{Output of bam-readcount, after reading in first 10 columns}

\item{vcf}{The VCF genotype file, to get the minor allele and subset to common SNPs. Only takes into account
so far minor alleles which are A,C,T or G (TODO: consider changing this)}
}
\value{
A data.table containing per SNP position: the total counts,
counts for the A,C,T and G alleles, the minor allele frequency estimate (b_estimate), the ALT allele,
and the SNP name in chromosome_position_refAllele (C_POS_REF) format
}
\description{
Transform the output of bam-readcount to data.table with minor allele frequency (MAF) estimate
Only works with bam-readcount output preprocessed to contain first 10 columns as detailed in vignette
}
