% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_weights.R
\name{estimate_weights}
\alias{estimate_weights}
\title{Estimate donor proportions (weights)}
\usage{
estimate_weights(b, A = gt, force_zero = TRUE)
}
\arguments{
\item{b}{a vector of minor allele frequencies with length k (k = number of donors)}

\item{A}{a genotype file with k+1 columns. k named donor columns containing minor allele dosages (0 for homozygotes for major allele,
1 for homozygotes of minor allele, 0.5 for heterozygotes).
Plus one extra column with "rn" name, containing SNP names. Number of rows of A must equal the length of b.}

\item{force_zero}{a logical. The least square regression result can return negative values that are very close to zero. If TRUE,
this parameter forces those values to zero. Defaults TRUE.}
}
\value{
a vector of proportions (w) with length k, where the sum equals 1.
}
\description{
Estimate donor proportions (weights) from the estimated MAF and the genotype
Solve for w in Aw = b using least square regression
With constraints sum(w) =1 and w can take values between 0-1.
}
