% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_weights.R
\name{is_identical_genotype}
\alias{is_identical_genotype}
\title{Mark rows where all columns (donors) in a genotype file have identical value}
\usage{
is_identical_genotype(x, ...)
}
\arguments{
\item{x}{Data table.}

\item{...}{Need to check what this is.}
}
\value{
Vector of row names where genotypes are identical
}
\description{
Checks per row if all columns have identical numeric value and marks those rows that are identical
}
